+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [Unreleased]
 
+### Added
+ - `IndexOf` methods.
+
 ## [0.9.4] - 2022-04-09
 
 ### Added